[#21] feat: fileToString, randomChipColor utility 추가
